pyproject.toml
[project]
name = "extropians"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.3",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
    "uvicorn>=0.41.0",
]
